added multi-tenant support, including routing

+++ backend/src/controllers/event.controller.ts
@@ -43,35 +43,52 @@ export class EventController {
           date: eventToCreate.date,
           targetClass: eventToCreate.targetClass,
           maxSeats: eventToCreate.maxSeats,
+          tenantId: (await UserService.currentTenant(req)).id,
         },
       });
       const logMessage = `Event ${event[0].id} erstellt`;
-      Log.build({ user: UserService.currentUser(req), message: logMessage }).save();
+      Log.log(UserService.currentTenant(req), UserService.currentUser(req), logMessage);
       res.status(200).send(event[0]);
     });
 
     app.put('/secure/events/:id/disabled/:disabled', async (req, res) => {
       const event = await Event.findByPk(req.params.id);
       if (event) {
-        event.disabled = req.params.disabled;
-        event.save();
-        const logMessage = `Event ${req.params.id} wurde auf ${event.disabled ? 'disabled' : 'enabled'} gesetzt`;
-        Log.build({ user: UserService.currentUser(req), message: logMessage }).save();
-        res.status(200).send(event);
+        // now check if the logged-in user belongs to the tenant that the event belongs to
+        if (event.tenantId === (await UserService.currentTenant(req)).id) {
+          event.disabled = req.params.disabled;
+          event.save();
+          const logMessage = `Event ${req.params.id} wurde auf ${event.disabled ? 'disabled' : 'enabled'} gesetzt`;
+          Log.log(UserService.currentTenant(req), UserService.currentUser(req), logMessage);
+          res.status(200).send(event);
+        } else {
+          res.status(403).send({ error: 'You are not authorized to delete this event' });
+        }
       } else {
-        res.sendStatus(404);
+        res.status(404).send({ error: 'Event not found' });
       }
     });
 
     app.delete('/secure/events/:id', async (req, res) => {
-      await Event.destroy({
-        where: {
-          id: req.params.id,
-        },
-      });
-      const logMessage = `Event ${req.params.id} gelöscht`;
-      Log.build({ user: UserService.currentUser(req), message: logMessage }).save();
-      res.status(200).send({ message: 'Event deleted' });
+      // first check if the event exists at all
+      const eventToDelete = (await Event.findByPk(req.params.id)) as Event;
+      if (eventToDelete) {
+        // now check if the logged-in user belongs to the tenant that the event belongs to
+        if (eventToDelete.tenantId === (await UserService.currentTenant(req)).id) {
+          await Event.destroy({
+            where: {
+              id: req.params.id,
+            },
+          });
+          const logMessage = `Event ${req.params.id} gelöscht`;
+          Log.log(UserService.currentTenant(req), UserService.currentUser(req), logMessage);
+          res.status(200).send({ message: 'Event deleted' });
+        } else {
+          res.status(403).send({ error: 'You are not authorized to delete this event' });
+        }
+      } else {
+        res.status(404).send({ error: 'Event not found' });
+      }
     });
 
     app.post('/events/:id/participant', async (req, res) => {
@@ -91,7 +108,7 @@ export class EventController {
             const createdParticipant = await Participant.build(newParticipant).save();
             res.status(200).send(createdParticipant);
           } else {
-            res.status(429).send({ message: 'All seats are taken already' });
+            res.status(409).send({ message: 'All seats are taken already' });
           }
         }
       } else {
